+++ converter/internal/prometheusconvert/validate.go
@@ -16,6 +16,7 @@ import (
 	prom_file "github.com/prometheus/prometheus/discovery/file"
 	prom_gce "github.com/prometheus/prometheus/discovery/gce"
 	_ "github.com/prometheus/prometheus/discovery/install" // Register Prometheus SDs
+	prom_ionos "github.com/prometheus/prometheus/discovery/ionos"
 	prom_kubernetes "github.com/prometheus/prometheus/discovery/kubernetes"
 	prom_docker "github.com/prometheus/prometheus/discovery/moby"
 )
@@ -94,6 +95,8 @@ func validateScrapeConfigs(scrapeConfigs []*prom_config.ScrapeConfig) diag.Diagn
 				diags.AddAll(validateDiscoveryKubernetes(sdc))
 			case *prom_aws.LightsailSDConfig:
 				diags.AddAll(validateDiscoveryLightsail(sdc))
+			case *prom_ionos.SDConfig:
+				diags.AddAll(validateDiscoveryIonos(sdc))
 			default:
 				diags.Add(diag.SeverityLevelError, fmt.Sprintf("unsupported service discovery %s was provided", serviceDiscoveryConfig.Name()))
 			}